Move cart addition functionality out of Item screen
