+++ app/src/screens/Item.tsx
@@ -1,27 +1,16 @@
 import React from 'react';
-import {
-	View,
-	Text,
-	TouchableOpacity,
-	ScrollView,
-	StyleSheet
-} from 'react-native';
-import { useDispatch } from 'react-redux';
+import { View, Text, ScrollView, StyleSheet } from 'react-native';
 import { useRoute, RouteProp } from '@react-navigation/native';
 import { useItemQuery } from '../types/api';
-import { Icon } from '../components/icons';
 import { AppStackParamList } from '../types/navigation';
-import { upsertItemToCart } from '../redux/carts/actions';
-import QuantityControl from '../components/item/QuantityControl';
 import ImageCarousel from '../components/item/ImageCarousel';
+import AddToCart from '../components/item/AddToCart';
 
 const Item = () => {
 	const { params } = useRoute<RouteProp<AppStackParamList, 'Item'>>();
-	const dispatch = useDispatch();
 	const [{ data }] = useItemQuery({
 		variables: { itemId: params.itemId }
 	});
-	const [quantity, setQuantity] = React.useState(0);
 
 	const item = data?.items_by_pk;
 
@@ -38,29 +27,10 @@ const Item = () => {
 					<Text style={{ fontSize: 18 }}>${item.unit_price}</Text>
 				</View>
 				<View style={styles.separator} />
-				<QuantityControl itemId={params.itemId} />
-				<TouchableOpacity
-					style={styles.cartButton}
-					onPress={() => {
-						dispatch(
-							upsertItemToCart({
-								storeId: item.store.id,
-								itemId: item.id,
-								quantity
-							})
-						);
-					}}
-				>
-					<Icon
-						size={22}
-						color='#FFFFFF'
-						name='plus'
-						style={{ marginRight: 5 }}
-					/>
-					<Text style={styles.buttonText}>Add to Cart</Text>
-				</TouchableOpacity>
+				<AddToCart storeId={item.store.id} itemId={item.id} />
 				<Text style={styles.descriptionHeader}>Description</Text>
 				<Text style={{ fontSize: 16 }}>{item?.description}</Text>
+				{/* Related Items */}
 			</View>
 		</ScrollView>
 	);
@@ -83,16 +53,6 @@ const styles = StyleSheet.create({
 		fontWeight: 'bold',
 		fontSize: 20
 	},
-	cartButton: {
-		marginVertical: 10,
-		width: '100%',
-		height: 40,
-		borderRadius: 8,
-		flexDirection: 'row',
-		justifyContent: 'center',
-		alignItems: 'center',
-		backgroundColor: '#202020'
-	},
 	metaContainer: {
 		flexDirection: 'row',
 		justifyContent: 'space-between',
@@ -110,11 +70,6 @@ const styles = StyleSheet.create({
 		height: 1,
 		backgroundColor: '#D3D3D3',
 		marginVertical: 10
-	},
-	buttonText: {
-		fontSize: 16,
-		fontWeight: '700',
-		color: '#FFFFFF'
 	}
 });
